requirements.txt
discord.py==2.4.0
discord.py[voice]==2.4.0
python-dotenv==1.0.1
yt-dlp==2025.01.26
PyNaCl==1.5.0
aiohttp==3.11.11
youtube-search~=2.1.2
validators==0.34.0
pytz==2023.3
Pillow==11.1.0
Flask==3.1.0
openai==1.61.1
dacite==1.8.1
PyYAML==6.0.2
certifi==2025.1.31
websockets==14.2
httpx==0.28.1
rich==13.9.4
prompt-toolkit==3.0.50
spacy==3.8.4
requests==2.32.3
beautifulsoup4~=4.12.2